requirements.txt
disnake
hashlib
os
typing
mysql-connector-python